:recycle: Custom Filter를 Bean으로 등록하여 싱글톤 패턴 적용
